Move Bitcoin members
