+++ src/NBitcoin.Tests/NetworkTests.cs
@@ -76,13 +76,13 @@ namespace NBitcoin.Tests
             Assert.Equal(Network.GetNetwork("mainnet"), network);
 
             Assert.Equal("Main", network.Name);
-            Assert.Equal(Network.BitcoinRootFolderName, network.RootFolderName);
-            Assert.Equal(Network.BitcoinDefaultConfigFilename, network.DefaultConfigFilename);
+            Assert.Equal(BitcoinMain.BitcoinRootFolderName, network.RootFolderName);
+            Assert.Equal(BitcoinMain.BitcoinDefaultConfigFilename, network.DefaultConfigFilename);
             Assert.Equal(0xD9B4BEF9, network.Magic);
             Assert.Equal(8333, network.DefaultPort);
             Assert.Equal(8332, network.RPCPort);
-            Assert.Equal(Network.BitcoinMaxTimeOffsetSeconds, network.MaxTimeOffsetSeconds);
-            Assert.Equal(Network.BitcoinDefaultMaxTipAgeInSeconds, network.MaxTipAge);
+            Assert.Equal(BitcoinMain.BitcoinMaxTimeOffsetSeconds, network.MaxTimeOffsetSeconds);
+            Assert.Equal(BitcoinMain.BitcoinDefaultMaxTipAgeInSeconds, network.MaxTipAge);
             Assert.Equal(1000, network.MinTxFee);
             Assert.Equal(20000, network.FallbackFee);
             Assert.Equal(1000, network.MinRelayTxFee);
@@ -158,13 +158,13 @@ namespace NBitcoin.Tests
             Assert.Empty(network.SeedNodes);
 
             Assert.Equal("TestNet", network.Name);
-            Assert.Equal(Network.BitcoinRootFolderName, network.RootFolderName);
-            Assert.Equal(Network.BitcoinDefaultConfigFilename, network.DefaultConfigFilename);
+            Assert.Equal(BitcoinMain.BitcoinRootFolderName, network.RootFolderName);
+            Assert.Equal(BitcoinMain.BitcoinDefaultConfigFilename, network.DefaultConfigFilename);
             Assert.Equal(0x0709110B.ToString(), network.Magic.ToString());
             Assert.Equal(18333, network.DefaultPort);
             Assert.Equal(18332, network.RPCPort);
-            Assert.Equal(Network.BitcoinMaxTimeOffsetSeconds, network.MaxTimeOffsetSeconds);
-            Assert.Equal(Network.BitcoinDefaultMaxTipAgeInSeconds, network.MaxTipAge);
+            Assert.Equal(BitcoinMain.BitcoinMaxTimeOffsetSeconds, network.MaxTimeOffsetSeconds);
+            Assert.Equal(BitcoinMain.BitcoinDefaultMaxTipAgeInSeconds, network.MaxTipAge);
             Assert.Equal(1000, network.MinTxFee);
             Assert.Equal(20000, network.FallbackFee);
             Assert.Equal(1000, network.MinRelayTxFee);
@@ -240,13 +240,13 @@ namespace NBitcoin.Tests
             Assert.Empty(network.SeedNodes);
 
             Assert.Equal("RegTest", network.Name);
-            Assert.Equal(Network.BitcoinRootFolderName, network.RootFolderName);
-            Assert.Equal(Network.BitcoinDefaultConfigFilename, network.DefaultConfigFilename);
+            Assert.Equal(BitcoinMain.BitcoinRootFolderName, network.RootFolderName);
+            Assert.Equal(BitcoinMain.BitcoinDefaultConfigFilename, network.DefaultConfigFilename);
             Assert.Equal(0xDAB5BFFA, network.Magic);
             Assert.Equal(18444, network.DefaultPort);
             Assert.Equal(18332, network.RPCPort);
-            Assert.Equal(Network.BitcoinMaxTimeOffsetSeconds, network.MaxTimeOffsetSeconds);
-            Assert.Equal(Network.BitcoinDefaultMaxTipAgeInSeconds, network.MaxTipAge);
+            Assert.Equal(BitcoinMain.BitcoinMaxTimeOffsetSeconds, network.MaxTimeOffsetSeconds);
+            Assert.Equal(BitcoinMain.BitcoinDefaultMaxTipAgeInSeconds, network.MaxTipAge);
             Assert.Equal(1000, network.MinTxFee);
             Assert.Equal(20000, network.FallbackFee);
             Assert.Equal(1000, network.MinRelayTxFee);
